Increase jwt ttl
This is not the right solution, but for now it will work. The problem is that we really don't know a good way to keep a user sessions if he just come back after the token in the local storage is expired. We can just ask him to log in again, but that don't seem right in a user experience point of view. If the user was loged in and did not pressed the logout button that means he just want to stay loged in.